Create Jacquelyn
Suggested Presentation Verbiage:

Team Paradigm’s prototype focuses on an “intuitive interface” that does not need much explaining; users can figure out how it works as soon as it is fired up.  

An intuitive interface is one that works the way the user expects it to; it must fit the user.

 Intuitive interfaces have real advantages:

•	Create happy users  
•	Require less training and support

If an interface works the way users expect it to, they do not need to learn how to use it.